backend/requirements.txt
asgiref==3.11.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
boto3==1.42.53
botocore==1.42.53
cbor2==5.8.0
certifi==2026.1.4
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
constantly==23.10.4
coverage==7.13.5
cryptography==46.0.5
daphne==4.2.1
Django==5.2.11
django[argon2]
django-redis==6.0.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.29.0
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
googleapis-common-protos==1.72.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.7
graphql-relay==3.2.0
httplib2==0.31.2
hyperlink==21.0.0
icalendar==7.0.1
idna==3.11
Incremental==24.11.0
inflection==0.5.1
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
msgpack==1.1.2
oauthlib==3.3.1
packaging==26.0
pillow==12.1.1
promise==2.3
proto-plus==1.27.1
protobuf==6.33.5
psycopg==3.3.2
psycopg-binary==3.3.2
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pyOpenSSL==25.3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
PyJWT==2.11.0
redis==7.1.1
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
resend==2.23.0
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.6.3
whitenoise==6.11.0
zope.interface==8.2